pyproject.toml
[project]
name = "hexlet-code"
version = "0.1.0"
description = "Page Analyzer project"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "flask>=3.0.0",
    "gunicorn>=23.0.0",
    "python-dotenv>=1.0.1",
]

[dependency-groups]
dev = [
    "ruff>=0.11.0",
]

[tool.ruff]
line-length = 88

[tool.ruff.lint]
select = ["E", "F", "I"]

[tool.hatch.build.targets.wheel]
packages = ["page_analyzer"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"